* Snippets overhauled to support subclassing them just to add an alternate page type with an alternate presentation, drawing from the same pool of content. This was always the intent but this is the first time we've tried to act on it and some issues were found.
Page types now have a separate '_typeCss' property used for the page settings dialog, which is specific to the page type, not the instance type; otherwise blog index pages and, let's say, press release index pages can't have separate page settings if they draw from the same pool of articles.

The first snippet page type with a given instance type is automatically registered as the "manager," the page type that provides the dropdown menu for actually creating snippets of that type. Any other page types for that instance type just have page settings and a dispatcher.

The type name, instance type, css name, typeCss name and manager status of a snippet page type are now passed from the server to the browser when pushing the call to the constructor.