One hundred per cent of editable fields are now in the schema.
A bc break for those who are already using snippetAllFields: you'll need to remove your manual calls for title, published, tags and hideTitle, because they are now included in snippetAllFields.

Of course you can use the options of snippetAllFields to render some things manually, or just choose to render everything manually.

There is support for a new "legacy" field in schema field definitions which allows an alternate field name to be checked. This allows us to gracefully handle existing overrides of new.html for events, people and anything else with hyphenated-fields ratherThanCamelCaseFields.